requirements.txt
aiohttp==3.7.4
asgiref==3.4.1
async-timeout==3.0.1
attrs==21.2.0
bson==0.5.10
cachetools==4.2.4
certifi==2021.10.8
cffi==1.15.0
chardet==3.0.4
charset-normalizer==2.0.7
click==8.0.3
cryptography==35.0.0
dj-database-url==0.5.0
Django==3.2.9
django-baton==2.2.3
django-cors-headers==3.7.0
django-environ==0.4.5
django-filter==21.1
django-heroku==0.3.1
django-leaflet==0.28.2
django-redis==5.0.0
django-rest-framework-roles==0.6
django-seconds-field==0.1.0
djangorestframework==3.12.2
djangorestframework-gis==0.17
djangorestframework-simplejwt==4.6.0
geographiclib==1.52
google-api-core==2.2.2
google-api-python-client==2.30.0
google-auth==2.3.3
google-auth-httplib2==0.1.0
googleapis-common-protos==1.53.0
gunicorn==20.1.0
httplib2==0.20.2
idna==3.3
install==1.3.4
drf-yasg
multidict==5.2.0
Pillow==8.2.0
protobuf==3.19.1
psycopg2==2.9.1
psycopg2-binary==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.0.1
pyOpenSSL==21.0.0
pyparsing==3.0.5
python-dateutil==2.8.2
python-decouple==3.5
pytz==2021.1
redis==3.5.3
requests==2.26.0
rsa==4.7.2
six==1.16.0
sqlparse==0.4.1
typing-extensions==4.0.1
uritemplate==4.1.1
urllib3==1.26.7
websockets==9.1
whitenoise==5.3.0
yarl==1.7.2
